// functions/api/_lib/analytics/types.ts
export type AnalyticsTimeframe =
  | 'last7Days'
  | 'last30Days'
  | 'last90Days'
  | 'yearToDate'
  | 'last12Months'
  | 'lifetime';
// ...
export type AnalyticsDateRange = {
  startDate: string;
  endDate: string;
};